Revert C++ changes so that mature-age is an int again as it should be. The corresponding fix to the data system to ensure rounding happens on that end will be in the next commit

+++ cvs/objects/ccarbon_model/include/land_carbon_densities.h
@@ -104,7 +104,7 @@ protected:
         DEFINE_VARIABLE( SIMPLE, "below-ground-carbon-density", mAvgBelowGroundCarbon, double ),
         
         //! Age at maturity.  This is used to grow forests slowly.
-        DEFINE_VARIABLE( SIMPLE, "mature-age", mMatureAge, /*int*/double )
+        DEFINE_VARIABLE( SIMPLE, "mature-age", mMatureAge, int )
     )
 };
